Codigo/Backend/huxgym/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
beautifulsoup4==4.9.3
cssutils==2.3.0
Django==3.2.3
django-inlinecss==0.3.0
django-naomi==0.8
djangorestframework==3.12.4
drf-base64==2.0
future==0.18.2
Pillow==8.2.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pynliner==0.8.0
python-decouple==3.4
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
django-cors-headers==3.7.0